Get weather data from BARRA-R2

Extract weather data for Australia from the BARRA-R2 weather data resource for a set of environments with defined latitude and longitude coordinates. BARRA-R2 data runs from Jan 1979 to Sept 2024.

Usage

get.BARRA.weather(
  Envs,
  Lats,
  Lons,
  Years,
  ncores = NULL,
  verbose = TRUE,
  dlprompt = FALSE
)

Arguments

Envs

Vector of environment names character strings.

Lats

Vector of latitude numeric values for each environment in the same order as Envs.

Lons

Vector of longitude numeric values for each environment in the same order as Envs.

Years

Vector of year integer values for each environment in the same order as Envs.

ncores

Number (integer) of cores to use for parallel processing of gridded data over muliple years. Use 1 to run in series. The default (NULL) will use the maximum available cores up to 5. If running in parallel, an output log text file will be created in the working directory.

verbose

Logical. Should progress be printed? Default = TRUE.

dlprompt

Logical. Should the user be prompted approve the total download size? Default = TRUE.

Value

A list of length 2:

$data is a list of matrices of weather data for each weather variable. Each data matrix has environment names as rows and days of the year as columns

$Env.info is a data frame of environment names and coordinate values for environments included in the data.

Details

Weather variables returned include:

daily_rain - Daily rainfall (mm)

max_temp - Maximum temperature (°C)

min_temp - Minimum temperature (°C)

vp_deficit - Vapour pressure deficit (hPa)

radiation - Solar exposure, consisting of both direct and diffuse components (MJ m-2)

day_lengths - Time between sunrise and sunset (h) not taken from BARRA-R2

VPD in hPa is calculated as \( VPD = 10(es - ea) \), where $$ es = 0.6108 \times \exp(\frac{17.27 \times T_{ave}}{T_{ave} + 237.3}) $$, \(T_{ave} \) is the mean temperature in °C, $$ ea = \frac{RH}{100} \times es $$, and \(RH\) is the relative humidity (%).

References

Su, C.H., Dharssi, I., Le Marshall, J., Le, T., Rennie, S., Smith, A., Stassen, C., Steinle, P., Torrance, J., Wang, C. and Warren, R.A., 2022. BARRA2: Development of the next-generation Australian regional atmospheric reanalysis. Bureau of Meteorology.

See also