Get weather data from CMIP6 QDC

Extract weather data from the Coupled Model Intercomparison Project phase 6 using a Quantile Delta Change method for Australia CMIP6 QDC weather data resource for a set of environments with defined latitude and longitude coordinates.

Usage

get.CMIP6.weather(
  Envs,
  Lats,
  Lons,
  Years,
  GCMs,
  SSPs,
  ncores = NULL,
  verbose = TRUE,
  dlprompt = FALSE
)

Arguments

Envs

Vector of environment names character strings.

Lats

Vector of latitude numeric values for each environment in the same order as Envs.

Lons

Vector of longitude numeric values for each environment in the same order as Envs.

Years

Vector of year integer values. Unlike get.SILO.weather() and get.BARRA.weather(), Years should not be the same length as Envs. Data for all locations in Envs will be extracted for all Years. Values must be within the possible time ranges of 1985-2014, 2035-2064, and 2070-2099.

GCMs

Vector of GCM names to get data from. For options see Details below.

SSPs

Vector of SSP names to get data from. For options see Details below.

ncores

Number (integer) of cores to use for parallel processing of gridded data. Use 1 to run in series. The default (NULL) will use the one less than the maximum available. If running in parallel, an output log text file will be created in the working directory.

verbose

Logical. Should progress be printed? Default = TRUE.

dlprompt

Logical. Should the user be prompted to approve the total download size? Default = TRUE.

Value

A multi-level list of $data and $Env.info weather variables within SSPs within GCMs:

$data is a list of matrices of weather data for each weather variable. Each data matrix has environment names as rows and days of the year as columns

$Env.info is a data frame of environment names and coordinate values for environments included in the data.

E.g. x$ACCESS-CM2$ssp585$daily_rain$data[1,1] will return the rainfall on the Jan 1st at the first location in the first year for the high emissions SSP585 for the ACCESS-CM2 GCM.

Details

The CMPI6 QDC dataset is hosted on the CSIRO data server includes climate projections for historical (1985-2014) and two future periods of years (2035-2064 and 2070-2099). All Years values must be within these time periods. The QDC method benchmarks CMIP6 against the observational data from the BARRA-R2 dataset so should be used in combination with the get.BARRA.weather() rather than the get.SILO.weather() function for observed environments that are outside of the 1985-2014 time period of the CMIP6 QDC historical time period.

Weather variables returned include:

daily_rain - Daily rainfall (mm)

max_temp - Maximum temperature (°C)

min_temp - Minimum temperature (°C)

vp_deficit - Vapour pressure deficit (hPa)

radiation - Solar exposure, consisting of both direct and diffuse components (MJ m-2)

day_lengths - Time between sunrise and sunset (h) not taken from CMIP6 QDC

VPD in hPa is calculated as \( VPD = 10(es - ea) \), where \( es = 0.6108 \times \exp(\frac{17.27 \times T_{ave}}{T_{ave} + 237.3}) \), \(T_{ave} \) is the mean temperature in °C, $$ ea = \frac{RH}{100} \times es $$, and \(RH\) is the relative humidity (%).

Possible options for GCMs (Global Climate Models):

ACCESS-CM2 - A much hotter future, and drier in most regions except the southeast

ACCESS-ESM1-5 - A hotter and much drier future

CMCC-ESM2 - A much hotter future with little change in mean rainfall (with regional exceptions)

CNRM-ESM2-1 - A much hotter future, much drier especially in the east, but wetter in the northwest

EC-Earth3 - A hotter and much wetter future for much of Australia (except southwest Western Australia)

MPI-ESM1-2-HR - Lower warming, mid-range changes in rainfall

NorESM2-MM - Lower warming, mid-range changes in rainfall

UKESM1-0-LL - Low probability, high impact case (high climate sensitivity, high Australian warming)

For further details see Grose et al. 2023

Possible options for SSPs (Shared Socio-economic Pathways) and equivalent Representative Concentration Pathways (RCP) with expected temperature increase range:

ssp126 - Sustainability (RCP2.6; 1.0-1.8°C)

ssp245 - Middle of the road (RCP4.5; 1.3-2.4°C)

ssp370 - Regional rivalry (No equivalent RCP; 2.8-4.6°C)

ssp585 - Fossil fueled development (RCP8.5; 3.3-5.7°C)

For further details see Riahi et al. 2017

An internet connection with high download speed is suggested for downloading gridded data for many environments. There may be problems with downloading files on a Mac OS.

References

Grose, M. R., Narsey, S., Trancoso, R., Mackallah, C., Delage, F., Dowdy, A., Di Virgilio, G., Watterson, I., Dobrohotoff, P., Rashid, H. A., Rauniyar, S., Henley, B., Thatcher, M., Syktus, J., Abramowitz, G., Evans, J. P., Su, C.-H., & Takbash, A. (2023). A CMIP6-based multi-model downscaling ensemble to underpin climate change services in Australia. Climate Services, 30, 100368.

Irving, D., & Macadam, I. (2024). Application-Ready Climate Projections from CMIP6 using the Quantile Delta Change method. CSIRO Climate Innovation Hub Technical Note 5. Technical Note.

Riahi, K., van Vuuren, D. P., Kriegler, E., Edmonds, J., O’Neill, B. C., Fujimori, S., Bauer, N., Calvin, K., Dellink, R., Fricko, O., Lutz, W., Popp, A., Cuaresma, J. C., Kc, S., Leimbach, M., Jiang, L., Kram, T., Rao, S., Emmerling, J., … Tavoni, M. (2017). The Shared Socioeconomic Pathways and their energy, land use, and greenhouse gas emissions implications: An overview. Global Environmental Change, 42, 153–168.

See also

Author

Nick Fradgley