Get environmental covariates from weather data

A function to calculate Environmental Covariates (ECs) from daily weather data such as derived from the get.SILO.weather() or get.BARRA.weather() functions.

Usage

get.W.ECs(
  weather,
  sow.dates,
  cardT = NULL,
  stg.TT = NULL,
  DTH.TT = NULL,
  verbose = TRUE
)

Arguments

weather

A list of length 2. Weather data as outputted from the get.SILO.weather().

weather$data is a list of data matrices for each covariate with rows as environments and days of the year as columns. Required weather covariate names for list items include:

daily_rain

max_temp

min_temp

vp_deficit

radiation

Additionally, if add.SMI() has also been use to estimate soil moisture index for the input weather object, them average SMI per stage ECs will also be calculated.

weather$Env.info is a data frame of info for each environment but is not required for this function.

sow.dates

Vector of character strings of dates of sowing for each trail environment in dd/mm/yyy format. Must be in the same order as the rownames of weather data in weather$data matrices.

cardT

Optional. Vector of minimum, optimal and maximum cardinal temperatures to calculate thermal time. Default values are min = 0, opt = 26, and max = 34. Custom values can be used to define other crop growth parameters.

stg.TT

Optional. Thermal time parameters that are used to estimate the number of days between wheat crop growth stages (see details). Default values:

Emergence to End of juvenile growth stages = 500

Heading to Flowering = 250

Flowering to Start of grain fill = 250

Start of Grain fill to End of grain fill = 250

End of grain fill to Maturity = 400

Custom values can be used to define other crop phenologies and growth rates

DTH.TT

Optional. Estimated thermal time from sowing that heading growth stage occurs. Default value is 1285.

verbose

Logical. Should progress be printed? Default if TRUE.

Value

A list of length 2:

$ECs - A data frame of weather EC values with environment names as rows and covariates as columns.

$gs.dates - A data frame of estimated dates in yyyy-mm-dd format of each growth stage per environment with environment names as rows and abbreviated growth stage names as columns.

Details

ECs are calculated for periods between crop growth stages that are estimated based on a thermal time degree days model defined by the cardT parameters. Crop growth stages abbreviations and equivalent Zadocks scale:

Sow - Sowing (GS0)

Emer - Emergence (GS10)

Juv - End of Juvenile (GS30)

He - Heading (GS55)

Flow - Flowering (GS65)

Sgf - Start of grain filling (GS71)

Egf - End of Grain filling (GS87)

Mat - Maturity (GS92)

Other abbreviations for ECs calculated between growth stage intervals and combined into EC names include:

Ndays - Number of days

TotRain - Total rainfall (mm)

Avtemp - Average of average daily temperatures (°C)

AvMintemp - Average of daily minimum temperatures (°C)

Avmaxtemp - Average of daily minimum temperatures (°C)

Ndays<0 - Number of frost days that the minimum temperature was below 0 °C

Ndays>26 - Number of warm days that max temp was over 26 °C

Ndays>34 - Number of hot days that max temp was over 34 °C

AveSR - Average solar radiation (MJ m-2)

AveVPD - Average vapour pressure deficit (hPa)

AvePQ - Average photothermal quotient (MJ m-2 day-1 °C-1)

AveDL - Average day length (hr)

AveSMI - Average Soil Moisture index (if add.SMI() function has been used)

For example, TotRain_He2Flw indicates the total rainfall between the heading and flowering growth stages.

Other specific ECs include:

TotRain_priorSow - The total rainfall between Jan 1st and the sowing day (mm)

Mintemp<0_Flw - Number number of frost days within 7 days of the estimated flowering date

For details of how ECs are calculated, see Fradgley et al. 2025.

References

Fradgley et al. (2025) Prediction of Australian wheat genotype by environment interactions and mega-environments, Under review.

Zadoks, J. C., Chang, T. T., & Konzak, C. F. (1974). A decimal code for the growth stages of cereals. Weed research, 14(6), 415-421.

See also